ajout de Product offer
Ajout d'une classe changement du pris en float

+++ src/fr/davidlegras/product/Product.java
@@ -1,11 +1,11 @@
 package fr.davidlegras.product;
 
 public abstract class Product {
-    private int price;
+    private float price;
     private String categorie;
     private String name;
 
-    public final int getPrice() {
+    public final float getPrice() {
         return price;
     }
     public final String getCategorie(){ return categorie;}
